Modularity of Create/Login/Forget
