+++ src/components/form/create-account/components/create-account-header.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import LandingHeader from "../../../landing/components/header/header";
+
+const Header: React.FC = () => {
+    return (
+        <div className="LandingHeader" data-aos="fade-down">
+            <LandingHeader />
+        </div>
+    );
+};
+
+export default Header;
