+++ src/components/form/forgot-password/components/forgot-password-header.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import LandingHeader from "../../../landing/components/header/header";
+
+const Header: React.FC = () => {
+    return (
+        <div data-aos="fade-down">
+            <LandingHeader />
+        </div>
+    );
+};
+
+export default Header;
